pyproject.toml
[project]
name = "airflow-dev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.1.7",
    "dbt-databricks>=1.11.5",
    "dbt-spark>=1.10.1",
]
